+++ src/main/java/com/buildingsmart/tech/ifcowl/ExpressReader.java
@@ -121,7 +121,7 @@ public class ExpressReader {
 
 	public ExpressReader(InputStream schemaInputStream) {
 		this.schemaInputStream = schemaInputStream;
-		Namespace.IFC = "http://ifcowl.openbimstandards.org/";
+		Namespace.IFC = "http://http://standards.buildingsmart.org/";
 	}
 
 	public void readAndBuild(){		
@@ -145,21 +145,39 @@ public class ExpressReader {
 		// else is accepted here
 		if (args.length != 2)
 			System.out
-			.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2' and 'IFC4' are accepted options");
+			.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1' and 'IFC4' are accepted options");
 		else {
 			String in = args[0];
 			if (in.equalsIgnoreCase("IFC2X3_Final")
 					|| in.equalsIgnoreCase("IFC2X3_TC1")
 					|| in.equalsIgnoreCase("IFC4_ADD1")
 					|| in.equalsIgnoreCase("IFC4_ADD2")
+					|| in.equalsIgnoreCase("IFC4_ADD2_TC1")
 					|| in.equalsIgnoreCase("IFC4x1_RC3")
 					|| in.equalsIgnoreCase("IFC4")) {
 				try {
 					InputStream instr = ExpressReader.class
 							.getResourceAsStream("/" + in + ".exp");
 					ExpressReader er = new ExpressReader(instr);
-					Namespace.IFC = "http://www.buildingsmart-tech.org/ifcOWL/"
-							+ in;
+					
+					String inAlt = in;
+					if (in.equalsIgnoreCase("IFC2X3_Final"))
+						inAlt = "IFC2X3/FINAL/";
+					if (in.equalsIgnoreCase("IFC2X3_TC1"))
+						inAlt = "IFC2X3/TC_1/";
+					if (in.equalsIgnoreCase("IFC4_ADD1"))
+						inAlt = "IFC4/ADD1/";
+					if (in.equalsIgnoreCase("IFC4_ADD2"))
+						inAlt = "IFC4/ADD2/";
+					if (in.equalsIgnoreCase("IFC4_ADD2_TC1"))
+						inAlt = "IFC4/ADD2_TC1/";
+					if (in.equalsIgnoreCase("IFC4x1_RC3"))
+						inAlt = "IFC4_1/";
+					if (in.equalsIgnoreCase("IFC4"))
+						inAlt = "IFC4/FINAL/";
+					
+					Namespace.IFC = "http://standards.buildingsmart.org/IFC/DEV/"
+							+ inAlt + "OWL";
 					er.readAndBuild();
 
 					er.outputEntitiesAndTypes(args[1], in);
@@ -176,7 +194,7 @@ public class ExpressReader {
 				}
 			} else
 				System.out
-				.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1' and 'IFC4' are accepted options");
+				.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1' and 'IFC4' are accepted options");
 		}
 	}