+++ src/main/java/com/buildingsmart/tech/ifcowl/ExpressReader.java
@@ -121,7 +121,7 @@ public class ExpressReader {
 
 	public ExpressReader(InputStream schemaInputStream) {
 		this.schemaInputStream = schemaInputStream;
-		Namespace.IFC = "http://http://standards.buildingsmart.org/";
+		Namespace.IFC = "http://standards.buildingsmart.org/";
 	}
 
 	public void readAndBuild(){		
@@ -145,7 +145,7 @@ public class ExpressReader {
 		// else is accepted here
 		if (args.length != 2)
 			System.out
-			.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1' and 'IFC4' are accepted options");
+			.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1', 'IFC4x1' and 'IFC4' are accepted options");
 		else {
 			String in = args[0];
 			if (in.equalsIgnoreCase("IFC2X3_Final")
@@ -153,7 +153,7 @@ public class ExpressReader {
 					|| in.equalsIgnoreCase("IFC4_ADD1")
 					|| in.equalsIgnoreCase("IFC4_ADD2")
 					|| in.equalsIgnoreCase("IFC4_ADD2_TC1")
-					|| in.equalsIgnoreCase("IFC4x1_RC3")
+					|| in.equalsIgnoreCase("IFC4x1")
 					|| in.equalsIgnoreCase("IFC4")) {
 				try {
 					InputStream instr = ExpressReader.class
@@ -164,14 +164,14 @@ public class ExpressReader {
 					if (in.equalsIgnoreCase("IFC2X3_Final"))
 						inAlt = "IFC2X3/FINAL/";
 					if (in.equalsIgnoreCase("IFC2X3_TC1"))
-						inAlt = "IFC2X3/TC_1/";
+						inAlt = "IFC2X3/TC1/";
 					if (in.equalsIgnoreCase("IFC4_ADD1"))
 						inAlt = "IFC4/ADD1/";
 					if (in.equalsIgnoreCase("IFC4_ADD2"))
 						inAlt = "IFC4/ADD2/";
 					if (in.equalsIgnoreCase("IFC4_ADD2_TC1"))
 						inAlt = "IFC4/ADD2_TC1/";
-					if (in.equalsIgnoreCase("IFC4x1_RC3"))
+					if (in.equalsIgnoreCase("IFC4x1"))
 						inAlt = "IFC4_1/";
 					if (in.equalsIgnoreCase("IFC4"))
 						inAlt = "IFC4/FINAL/";
@@ -194,7 +194,7 @@ public class ExpressReader {
 				}
 			} else
 				System.out
-				.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1' and 'IFC4' are accepted options");
+				.println("Usage: java ExpressReader expressSchemaname pathToOutputFile \nExample: java ExpressReader IFC2X3_TC1 C:/outputfile.owl \nNote: only 'IFC2X3_Final', 'IFC2X3_TC1', 'IFC4_ADD1', 'IFC4_ADD2', 'IFC4_ADD2_TC1', 'IFC4x1' and 'IFC4' are accepted options");
 		}
 	}