+++ CHANGES.md
@@ -3,6 +3,7 @@ EXPRESStoOWL Change Log
 Current Development 0.3-SNAPSHOT
 - changed namespace and output to http://standards.buildingsmart.org/IFC/DEV/
 - added conversion and files for IFC_ADD2 and IFC_ADD2_TC1
+- added conversion and files for IFC4x1
 
 Release 0.2 (2018-01-04)
 - Added additional ontology files