+++ CHANGES.md
@@ -1,6 +1,6 @@
 EXPRESStoOWL Change Log
 
-Current Development 0.3-SNAPSHOT
+Release 0.3 (2019-07-31)
 - changed namespace and output to http://standards.buildingsmart.org/IFC/DEV/
 - added conversion and files for IFC_ADD2 and IFC_ADD2_TC1
 - added conversion and files for IFC4x1